Remove gemoji submodule and replace with shallow copy of file
Cloning literally hundreds of MBs for a single json file just was too
inconvenient for me. Plus, every time the upstream would change there
would still be manual effort necessary to pull the changes in. Just instead
of juggling a submodule, now the effort is to just download a new version
of the file.